+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## [Versão 3.77.132]
+
+## [Versão 3.78.132]
+- Adicionado campo para a assinatura do diretor 
+
+## [Versão 3.77.131]
 - Adicionado unidades de medida para o lançamento no estoque
 
 ## [Versão 3.77.130]
@@ -9,7 +13,6 @@
 
 ## [Versão 3.76.129]
 - Correção na descrição da séries para as inconsistências sagres
-
 
 ## [Versão 3.76.128]
 - Corrigido erro que não mostrava o nome da escola e das turmas nas inconsistências sagres